requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
pikepdf==8.10.1
Werkzeug==3.0.1
